# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses",
        "operationId": "list_courses",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "srch_tra_st_dt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "\ud6c8\ub828\uc2dc\uc791\uc77c From (YYYYMMDD \ub610\ub294 YYYY-MM-DD)",
              "title": "Srch Tra St Dt"
            },
            "description": "\ud6c8\ub828\uc2dc\uc791\uc77c From (YYYYMMDD \ub610\ub294 YYYY-MM-DD)"
          },
          {
            "name": "srch_tra_end_dt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "\ud6c8\ub828\uc2dc\uc791\uc77c To (YYYYMMDD \ub610\ub294 YYYY-MM-DD)",
              "title": "Srch Tra End Dt"
            },
            "description": "\ud6c8\ub828\uc2dc\uc791\uc77c To (YYYYMMDD \ub610\ub294 YYYY-MM-DD)"
          },
          {
            "name": "srch_tra_organ_nm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\ud6c8\ub828\uae30\uad00\uba85",
              "title": "Srch Tra Organ Nm"
            },
            "description": "\ud6c8\ub828\uae30\uad00\uba85"
          },
          {
            "name": "srch_tra_process_nm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\ud6c8\ub828\uacfc\uc815\uba85",
              "title": "Srch Tra Process Nm"
            },
            "description": "\ud6c8\ub828\uacfc\uc815\uba85"
          },
          {
            "name": "has_reg_course_man",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "\uc218\uac15\uc2e0\uccad \uc778\uc6d0 \uc788\uc74c (regCourseMan > 0)",
              "default": false,
              "title": "Has Reg Course Man"
            },
            "description": "\uc218\uac15\uc2e0\uccad \uc778\uc6d0 \uc788\uc74c (regCourseMan > 0)"
          },
          {
            "name": "page_num",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page Num"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/owned-search": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Search Owned Courses",
        "operationId": "search_owned_courses",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 2023,
              "description": "\ud6c8\ub828\uc2dc\uc791\uc77c \uae30\uc900 \uc870\ud68c \ub144\ub3c4",
              "title": "Year"
            },
            "description": "\ud6c8\ub828\uc2dc\uc791\uc77c \uae30\uc900 \uc870\ud68c \ub144\ub3c4"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0.0,
              "description": "\uad00\ub828\ub3c4 \uc784\uacc4\uce58 (0\uc774\uba74 \ubbf8\uc801\uc6a9)",
              "default": 0,
              "title": "Min Score"
            },
            "description": "\uad00\ub828\ub3c4 \uc784\uacc4\uce58 (0\uc774\uba74 \ubbf8\uc801\uc6a9)"
          },
          {
            "name": "has_reg_course_man",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "\uc218\uac15\uc2e0\uccad \uc778\uc6d0 \uc788\uc74c (regCourseMan > 0)",
              "default": false,
              "title": "Has Reg Course Man"
            },
            "description": "\uc218\uac15\uc2e0\uccad \uc778\uc6d0 \uc788\uc74c (regCourseMan > 0)"
          },
          {
            "name": "page_num",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page Num"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/courses/search": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Search Courses",
        "operationId": "search_courses",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Keyword"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 5,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseSearchHit"
                  },
                  "title": "Response Courses-Search Courses"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/index": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Index Courses From Work24",
        "operationId": "index_courses_from_work24",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseIndexResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/courses/legacy-index": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Enqueue Legacy Course Index",
        "operationId": "enqueue_legacy_course_index",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyIndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyIndexResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CourseListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "page_num": {
            "type": "integer",
            "title": "Page Num"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "page_num",
          "page_size"
        ],
        "title": "CourseListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}